Create Reverse Array
Write a recursive function to implement the recursive algorithm to reverse the elements of an
array between two indices. Also, write a program to test your function. Your program should
prompt the user for the integers in the array, as well as the low index and then the high index.