Create smallestLargest
Write a recursive function that returns both the smallest and the largest element in
an int array. Also, write a program to test your function. This program should ask the user for an
array of int and should output the smallest and the largest element to the console in the following
format: