Create maze
Create a class called "Maze" that includes a recursive call to find a path through a given text maze. The first 2 lines in the text file represent the size of the maze and the the next 2 lines represent the starting location in the maze. A sample text file is attached.

I will test your program on a similar text file. 

maze.txt

1.) You will first have to set up your Scanner to the maze text file.

2.) You must dynamically create a 2-dimensional character array.(remember there is no .nextChar() )

3.) Write a recursive method to "find your way" through the maze. (Each recursive call will represent one move in the maze)